+++ module/reputation.lua
@@ -1,8 +1,9 @@
 
 local realm = GetRealmName();
 local char = UnitName('player');
+local _, ns = ...;
 
-function updateReputation ()
+function ns:updateReputation ()
 
 	local hasFilter = false;